update sales order id

+++ src/main/java/com/xsis/mp1/model/SalesOrder.java
@@ -72,7 +72,7 @@ public class SalesOrder {
 		this.id = id;
 	}
 
-	public double getGrandTotal() {
+	public float getGrandTotal() {
 		return grandTotal;
 	}
